+++ packages/modules/imodel-browser/CHANGELOG.md
@@ -1,6 +1,34 @@
 # Change Log - @itwin/imodel-browser-react
 
-This log was last generated on Fri, 06 Jun 2025 18:58:31 GMT and should not be manually modified.
+This log was last generated on Tue, 15 Jul 2025 19:19:08 GMT and should not be manually modified.
+
+## 3.2.0
+Tue, 15 Jul 2025 19:19:08 GMT
+
+### Minor changes
+
+- Add cellOverrides option to iModelGrid and iTwinGrid
+
+## 3.1.2
+Thu, 10 Jul 2025 23:35:54 GMT
+
+### Patches
+
+- Fix iModel opening when opening action dropdown
+
+## 3.1.1
+Thu, 10 Jul 2025 16:59:48 GMT
+
+### Patches
+
+- Refactor IModelTile component to allow hovering over name label
+
+## 3.1.0
+Fri, 27 Jun 2025 06:46:50 GMT
+
+### Minor changes
+
+- Add getBadge property on iModelTile in order to set badge based on iModel props
 
 ## 3.0.7
 Fri, 06 Jun 2025 18:58:31 GMT